update to import dimelo as dm
