+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
 ignore = E203, E266, E501, W503
-exclude = .git,docs
+exclude = .git,docs,__init__.py
