+++ dimelo/__init__.py
@@ -10,3 +10,6 @@ dimelo allows you to perform quality control and plot modified bases from bam fi
 .. automodule:: dimelo.plot
 
 """
+from .browser_sm_roi import browser_sm_roi
+from .enrich_sm_roi import enrich_sm_roi
+from .parse_bam import parse_bam
